Refactor version check code
